* commit '9b0bd2b32061e852c3233f6784e974b5226c5821':
  Add casts to avoid build warnings with gcc-4.7

+++ keystore/keystore_client.cpp
@@ -47,7 +47,7 @@ ResponseCode keystore_cmd(command_code_t cmd, Keystore_Reply* reply, int numArgs
             return SYSTEM_ERROR;
         }
 
-        uint8_t bytes[2] = { argLen >> 8, argLen };
+        uint8_t bytes[2] = { (uint8_t)(argLen >> 8), (uint8_t)argLen };
         if (TEMP_FAILURE_RETRY(send(sock, bytes, 2, MSG_NOSIGNAL)) != 2
                 || TEMP_FAILURE_RETRY(send(sock, arg, argLen, MSG_NOSIGNAL))
                         != static_cast<ssize_t>(argLen)) {
